Fixed T-Poses when crouching with SMGs and Pistols
I've seen a lot of people complain about this and it also affected me. It turns out the animation codes for some of the animations were incorrect.
ACT_COVER_PISTOL_LOW and
ACT_COVER_SMG1_LOW as well as
ACT_RANGE_AIM_PISTOL_LOW
do not exist and have been replaced with the correct (or at least most fitting) counterparts.